Fixed login without nativeAuth token
