+++ src/controllers/TransactionsHistoryController/TransactionsHistoryController.ts
@@ -1,7 +1,7 @@
 import { IGetHistoricalTransactionsParams } from 'types/transaction-list-item.types';
 import { getTransactionsHistory } from 'utils/transactions';
 
-// Note: Will replace TransactionsTableController in the future
+// TODO: Will replace TransactionsTableController in the future
 export const TransactionsHistoryController = {
   async getTransactionsHistory(params: IGetHistoricalTransactionsParams) {
     const transactions = await getTransactionsHistory(params);
