+++ src/core/providers/DappProvider/helpers/login/helpers/getAccountFromToken.ts
@@ -1,9 +1,7 @@
-import { fetchAccount } from 'utils/account/fetchAccount';
 import { getModifiedLoginToken } from './getModifiedLoginToken';
 
 interface GetAccountFromTokenType {
   address: string;
-  apiAddress: string;
   originalLoginToken?: string;
   extraInfoData: {
     multisig?: string;
@@ -14,7 +12,6 @@ interface GetAccountFromTokenType {
 export async function getAccountFromToken({
   originalLoginToken,
   extraInfoData,
-  apiAddress,
   address
 }: GetAccountFromTokenType) {
   const modifiedLoginToken = await getModifiedLoginToken({
@@ -27,13 +24,7 @@ export async function getAccountFromToken({
 
   const accountAddress = modifiedLoginToken != null ? tokenAddress : address;
 
-  const account = await fetchAccount({
-    address: accountAddress,
-    baseURL: apiAddress
-  });
-
   return {
-    account,
     address: accountAddress,
     modifiedLoginToken
   };
